pyproject.toml
[project]
name = "ai-groq-economics-toolkit"
version = "0.1.0"
description = "A cost modeling and unit economics toolkit for Groq workloads."
authors = [
    { name = "Ron Smith Jr" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pyyaml>=6.0",
    "tabulate>=0.9.0"
]

[project.scripts]
groq-econ = "groq_econ.cli:main"

[tool.setuptools.packages.find]
where = ["."]